add visualization code + 'standalone' hops version
I have used this code for creating the animations in the presentation.
As a side effect, there now exists a HOPS version that gives you, among
visualization-specific stuff, a map of tree vertices to graph vertices
that specifies the embedding that was found (if any).

Might be useful to some.